fix trait definition and move it to separate file
