+++ raml/samples/resource_types.raml
@@ -49,14 +49,16 @@ resourceTypes:
            description: Return <<resourcePathName>> that have their <<queryParamName>> matching the given value
          <<fallbackParamName>>:
            description: If no values match the value given for <<queryParamName>>, use <<fallbackParamName>> instead
-#traits:
-#  - secured:
-#      description: Some requests require authentication
-#      queryParameters:
-#        <<methodName>>: # e.g. get:
-#          description: A <<methodName>>-token pair is required  # e.g. A get-token pair...
-#          example: <<methodName>>=h8duh3uhhu38   # e.g. get=h8duh3uhhu38
-
+traits:
+  - secured:
+     description: requests to <<methodName>> require authentication
+     queryParameters:
+       <<tokenName>>:
+         description: A valid <<tokenName>> is required
+  - paged:
+      queryParameters:
+        numPages:
+          description: The number of pages to return, not to exceed <<maxPages>>
 
 /Users:
   type: collection
@@ -81,4 +83,4 @@ resourceTypes:
   type: { searchableCollection: { queryParamName: title, fallbackParamName: digest_all_fields } }
   get:
     description : get all books
-    # is: [ secured: { tokenName: access_token }, paged: { maxPages: 10 } ]
+    is: [ secured: { tokenName: access_token }, paged: { maxPages: 10 } ]
