+++ raml/samples/simple_example.raml
@@ -46,11 +46,11 @@ uriParameters:
 baseUri: /lala
 protocols: [HTTP, HTTPS]
 traits:
-    secured:
+  - secured:
       queryParameters:
         <<tokenName>>:
           description: A valid <<tokenName>> is required
-    paged:
+  - paged:
       queryParameters:
         numPages:
           description: The number of pages to return, not to exceed <<maxPages>>
